+++ src/felupe/mesh/_mesh.py
@@ -559,6 +559,9 @@ class Mesh(DiscreteGeometry):
           Number of cells:
             quad: 100
             quad: 100
+        
+        ..  image:: images/mesh_container.png
+            :width: 400px
 
         The meshes of the mesh container are :func:`stacked <felupe.mesh.stack>`.
 
@@ -573,10 +576,16 @@ class Mesh(DiscreteGeometry):
         the number of cells is unchanged.
 
         >>> mesh = fem.mesh.sweep(stack)
+        >>> mesh
         <felupe Mesh object>
           Number of points: 220
           Number of cells:
             quad: 200
+        
+        >>> ax = mesh.imshow(opacity=0.6)
+        
+        ..  image:: images/mesh_sweep.png
+            :width: 400px
 
         ..  note::
             The :class:`~felupe.MeshContainer` may be directly created with